+++ src/templates/Home/Addresses/index.ts
@@ -1,0 +1,1 @@
+export {Addresses} from './Addresses'
